% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{auc}
\alias{auc}
\title{Calculate area under the curve for two vectors}
\usage{
auc(x, y, zero = 0)
}
\arguments{
\item{x}{a vector of x values}

\item{y}{a vector of y values}

\item{zero}{which value to consider as the y floor. defaults to 0.}
}
\value{
a dataframe containing the y and x values, as well as the calculated
auc for each pair of points and the cumulative auc as X progresses.
}
\description{
Calculate area under the curve for two vectors
}
