% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyro_functions.R
\name{aux_fun_process_NAs}
\alias{aux_fun_process_NAs}
\title{auxiliary function of patch_NAs}
\usage{
aux_fun_process_NAs(nas, col, input, patch_method, verbose)
}
\arguments{
\item{nas}{A table with the start and stop rows of the NA intervals}

\item{col}{the column being patched}

\item{input}{A dataframe containing Timestamps on the first column and
matching data on the remaining. If input contains a "phase" column, it will
be transported to the output unchanged.}

\item{patch_method}{One of:
'linear' to capture the nearest before and after non-NA values and make a
linear interpolation.
'before' to find the nearest non-NA value before the NA and use it to fill
the gap.
'after'  to do the same as above but with the nearest value coming after
the NA.}

\item{verbose}{Logical. Should exceptions be reported out loud.
Defaults to TRUE.}
}
\description{
Internal. Performs the actual patching for one column at a time.
}
\keyword{internal}
