% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epoc_functions.R
\name{calc_auc}
\alias{calc_auc}
\title{Calculate area under the curve}
\usage{
calc_auc(mmr, smr, smr_col, smr_buffer = 0.1)
}
\arguments{
\item{mmr}{post-chasing resp object}

\item{smr}{pre-chasing resp object}

\item{smr_col}{name of the SMR column to use (depends on the methods used
to calculate the SMR in the SMR object)}

\item{smr_buffer}{Defaults to 0.1. Defines the zero for the auc calculations.
0.1 means the AUC calculation will be 0 once the MO2 measured post-chase
returns to 1.1 times the respective SMR of the animal.}
}
\value{
an updated mmr object, now containing 1) an updated mr object which
shows the delta MO2, and 2) a new auc object. The auc object contains both
the auc between each pair of points, and also the cumulative auc.
}
\description{
Calculate area under the curve
}
