% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_functions.R
\name{calc_linear_bg}
\alias{calc_linear_bg}
\title{Internal function}
\usage{
calc_linear_bg(input, pre, post)
}
\arguments{
\item{input}{A list containing trimmed oxygen data.
Obtained through \code{\link{process_experiment}}.}

\item{pre}{(optional) A data frame containing pre-test background readings.
Obtained through \code{\link{calc_bg}}.}

\item{post}{(optional) A data frame containing post-test background readings.
Obtained through \code{\link{calc_bg}}.}
}
\description{
Calculates the linear progression between pre- and post-background readings,
used by \code{\link{subtract_bg}}
}
\keyword{internal}
