% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_functions.R
\name{calc_mr}
\alias{calc_mr}
\title{Calculate metabolic rates}
\usage{
calc_mr(input)
}
\arguments{
\item{input}{a data frame with columns slope_cor (mandatory),
water_vol (mandatory), and animal_mass (optional).}
}
\value{
The input with columns mr_abs and mr_g if animal_mass provided.
}
\description{
Calculates absolute and mass-specific metabolic rates.
}
