% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_functions.R
\name{calc_single_slope}
\alias{calc_single_slope}
\title{Calculate slopes from clean measurements}
\usage{
calc_single_slope(input, probe, cycle, skip = 0, max_duration = Inf)
}
\arguments{
\item{input}{A resp object with a trimmed table.}

\item{probe}{The probe for which to calculate a slope}

\item{cycle}{The cycle for which to calculate a slope}

\item{skip}{Number of seconds to skip at the start
of the phase.}

\item{max_duration}{Only use points up to this time
into the phase (in seconds). Defaults to Inf.}
}
\description{
Calculate slopes from clean measurements
}
