% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smr_functions.R
\name{calc_smr}
\alias{calc_smr}
\title{Calculate SMR using various methods}
\usage{
calc_smr(mr, G = 1:4, q = c(0.2, 0.25), p = 0.1, n = 10)
}
\arguments{
\item{mr}{A data frame containing the metabolic rates.
The output of \code{\link{calc_mr}}.}

\item{G}{A vector of the number of clusters to use for MLND SMR estimations.
Defaults to 1:4. Set to NULL to skip MLND method.}

\item{q}{A vector of the quantiles to use for the quantile SMR estimations.
Defaults to 0.2 and 0.25. Set to NULL to skip the quantile method.}

\item{p}{A vector of the percentages to use for low percent SMR estimations.
Percent methods discard the five lowest measurements before calculating the
average of the lowest x percentage. Defaults to 0.1. Set to NULL to skip
the percent method.}

\item{n}{A vector of the number of measurements to use for the low SMR
estimations. This method grabs the lowest n measurements and averages them.
No values are discarded using this method. Defaults to 10. Set to NULL to
skip the low n method.}
}
\description{
Calculate SMR using various methods
}
\references{
{Chabot, D., Steffensen, J. F., & Farrell, A. P. (2016).
The determination of standard metabolic rate in fishes.
Journal of Fish Biology, 88(1), 81-121.}
}
