% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o2_calcs_functions.R
\name{calculate.bg}
\alias{calculate.bg}
\title{dummy doc}
\usage{
calculate.bg(
  input,
  O2_col,
  method = c("mean", "first", "last"),
  force.linear = TRUE,
  smoothing = 30
)
}
\description{
dummy doc
}
