% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o2_calcs_functions.R
\name{calculate.MR}
\alias{calculate.MR}
\title{Calculation of Metabolic Rate}
\usage{
calculate.MR(slope.data, density = 1000)
}
\arguments{
\item{slope.data}{a data frame obtained by using the function \code{\link{extract.slope}}}

\item{density}{numeric: the density of an animal body (\eqn{kg/m^{3}})}
}
\value{
The function returns a data frame with calculated background respiration, absolute and mass-specific metabolic rates.
}
\description{
The function is used to calculate and plot background respiration, absolute and mass-specific metabolic rates.
}
