% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o2_calcs_functions.R
\name{calculate_linear_bg_progression}
\alias{calculate_linear_bg_progression}
\title{Internal function}
\usage{
calculate_linear_bg_progression(pre.bg, post.bg, meas.data)
}
\description{
Internal function
}
\keyword{internal}
