% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyroscience_functions.R
\name{check.pyrocience.data}
\alias{check.pyrocience.data}
\title{Confirm that the pyroscience input contains the expected columns}
\usage{
check.pyrocience.data(x)
}
\arguments{
\item{x}{the input data.frame}
}
\value{
a data frame with ordered columns
}
\description{
Confirm that the pyroscience input contains the expected columns
}
\keyword{internal}
