% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{check_arg_in_data}
\alias{check_arg_in_data}
\title{check that the requested argument value is present in the data}
\usage{
check_arg_in_data(arg, data, name, verbose = TRUE)
}
\arguments{
\item{arg}{the argument value}

\item{data}{A vector of the target data}

\item{name}{For messaging purposes; the type of data we're working with.}

\item{verbose}{Should a warning be sent if the argument does not fully
match the data?}
}
\value{
the matching elements of arg
}
\description{
check that the requested argument value is present in the data
}
\keyword{internal}
