% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases_functions.R
\name{check_devices_match}
\alias{check_devices_match}
\title{confirm that the devices listed in the pyro input
are present in the phases input}
\usage{
check_devices_match(pyro_names, phase_names)
}
\arguments{
\item{pyro_names}{A vector of column names from the compiled_data object}

\item{phase_names}{the names of the phases list}
}
\value{
nothing. Used for side effects
}
\description{
confirm that the devices listed in the pyro input
are present in the phases input
}
\keyword{internal}
