% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases_functions.R
\name{check_phases}
\alias{check_phases}
\title{Check if phases line up correctly and are not repeated}
\usage{
check_phases(phases, fix_phases = FALSE)
}
\arguments{
\item{phases}{The phases list. The "phases" object in the list generated
by \code{\link{load_experiment}}}

\item{fix_phases}{Automatically correct any overlaps, gaps, or repeated
phase names found in the file.}
}
\value{
The phases list, with troublesome rows saved to the attributes of
each probe's data frame.
}
\description{
Check if phases line up correctly and are not repeated
}
