% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases_functions.R
\name{check_probes_match}
\alias{check_probes_match}
\title{confirm that the probes listed for each device in the pyro input
are present in the phases input}
\usage{
check_probes_match(pyro_names, phases)
}
\arguments{
\item{pyro_names}{A vector of column names from the compiled_data object}

\item{phases}{The phases list.}
}
\value{
nothing. Used for side effects
}
\description{
confirm that the probes listed for each device in the pyro input
are present in the phases input
}
\keyword{internal}
