% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o2_calcs_functions.R
\name{clean_meas}
\alias{clean_meas}
\title{Prepare imported measurements for further analyses}
\usage{
clean_meas(input, wait = 0, auto_cut_last = FALSE)
}
\arguments{
\item{input}{the data frame containing imported oxygen measurements}

\item{wait}{integer: the number of first rows for each measurement phase (M)
which should be reassigned to the wait phase (W). Note: If your
phase-tracking device already assigns a wait phase, set this to 0.}

\item{auto_cut_last}{Should the last recorded phase be automatically cut?
useful if the experiment was terminated mid-measurement.}
}
\value{
A data frame containing measurements valid for further analyses
}
\description{
Prepare imported measurements for further analyses
}
