% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyro_functions.R
\name{compile_pyro_data}
\alias{compile_pyro_data}
\title{dummy documentation}
\usage{
compile_pyro_data(folder, legacy.device.name = TRUE)
}
\arguments{
\item{folder}{the pyroscience run folder, containing a "ChannelData" folder inside}

\item{pattern}{The type of variable to look for. Currently accepted values: "Oxygen", "pH"}
}
\description{
scan a target folder for pyroscience files matching a pattern and combine them.
}
