% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o2_calcs_functions.R
\name{compile_rolling_mmr}
\alias{compile_rolling_mmr}
\title{Wrapper of fraction_mr for MMR cycles}
\usage{
compile_rolling_mmr(input, smoothing)
}
\arguments{
\item{input}{The output of process_pyro_mr}

\item{smoothing}{How many points should be gathered for each calculation of the rolling MR}
}
\description{
Wrapper of fraction_mr for MMR cycles
}
