% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{compile_sources}
\alias{compile_sources}
\title{merge a list of data sources into a single table using date_time
as the link. Ensures there is one observation per second.}
\usage{
compile_sources(input)
}
\arguments{
\item{input}{a list of equally formatted tables}
}
\value{
A table with the compiled data
}
\description{
merge a list of data sources into a single table using date_time
as the link. Ensures there is one observation per second.
}
