% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{conv_w_to_ml}
\alias{conv_w_to_ml}
\title{convert weight to volume in ml}
\usage{
conv_w_to_ml(w, density = 1)
}
\arguments{
\item{w}{a units object in either grams or kilograms}

\item{density}{The density. defaults to 1 (1 g = 1 ml).
Higher densities lead to lower volumes, e.g. if density = 2,
then 1 g = 0.5 ml.}
}
\value{
a units object in ml
}
\description{
convert weight to volume in ml
}
