% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyro_functions.R
\name{discard_cycles}
\alias{discard_cycles}
\title{Discard readings}
\usage{
discard_cycles(input, probes, cycles)
}
\arguments{
\item{input}{A computer-friendly data frame.
The output of \code{\link{melt_resp}} or any downstream function.}

\item{probes}{The probe(s) from which to discard data. Ommit to discard phases
from all probes.}

\item{cycles}{The cycles(s) to discard.}
}
\value{
the input data frame without the discarded readings.
}
\description{
Discard the data from one or more phases of one or more probes.
}
