% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyro_functions.R
\name{discard_phase}
\alias{discard_phase}
\title{Discard readings}
\usage{
discard_phase(input, probe, phase)
}
\arguments{
\item{input}{A computer-friendly data frame.
The output of \code{\link{melt_resp}} or any downstream function.}

\item{probe}{The probe(s) from which to discard data. Ommit to discard phases
from all probes.}

\item{phase}{The phase(s) to discard.}
}
\value{
the input data frame without the discarded readings.
}
\description{
Discard the data from one or more phases of one or more probes.
}
