% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o2_calcs_functions.R
\name{exclude_rolling_mmr_segment}
\alias{exclude_rolling_mmr_segment}
\title{Remove part of the mr trace}
\usage{
exclude_rolling_mmr_segment(rolling_mmr, probe, start, end)
}
\arguments{
\item{rolling_mmr}{the output of compile_rolling_mmr}

\item{probe}{which probe to act upon}

\item{start}{second from which to start exclusion}

\item{end}{second where to end exclusion}
}
\description{
Useful when there are artifacts along the cycle that could disrupt fractioned mr calculation.
}
