% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_functions.R
\name{exclude_rolling_mr_segment}
\alias{exclude_rolling_mr_segment}
\title{Remove part of the mr trace}
\usage{
exclude_rolling_mr_segment(input, probe, cycle, from, to)
}
\arguments{
\item{input}{An experiment list containing
the output of \code{\link{roll_mr}}.}

\item{probe}{which probe to act upon}

\item{cycle}{Which phase to act upon}

\item{from}{second from which to start exclusion}

\item{to}{second where to end exclusion}
}
\description{
Useful when there are artifacts along the cycle that could disrupt fractioned
mr calculation. Refinds max mmr after removing the desired segment.
}
