% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmr_functions.R
\name{extract_mmr}
\alias{extract_mmr}
\title{Find cycle with highest mr for each probe}
\usage{
extract_mmr(mr)
}
\arguments{
\item{mr}{The metabolic rate data frame. The output of \code{\link{calc_mr}}.}
}
\value{
A data frame with the mmr cycle and and MO2 for each probe.
}
\description{
Find cycle with highest mr for each probe
}
