% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_functions.R
\name{extrapolate_bg}
\alias{extrapolate_bg}
\title{Expand background readings linearly}
\usage{
extrapolate_bg(input, to)
}
\arguments{
\item{input}{an experiment list containing background readings.
The output of \code{\link{calc_bg}}.}

\item{to}{The number of seconds to which to extend the background readings.}
}
\value{
the updated input.
}
\description{
Use this if your measurement phase is longer than your background and you
want ot expand your background linearly to accomodate the full duration
of the phase.
}
