% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FishResp.R
\docType{package}
\name{FishResp}
\alias{FishResp}
\title{FishResp: Analytical Tool for Aquatic Respirometry}
\description{
THIS IS A FORKED VERSION OF THE FishResp PACKAGE!
}
\details{
Do not use this repository if you are looking for the real FishResp!

Go here instead: https://github.com/embedded-sergey/FishResp-Rpackage
}
\references{
{
Before using the R package 'FishResp', we recommend reading the publication describing this software and two keystone reviews devoted to metabolic rate measurements of fish using an intermittent-flow approach:
\enumerate{
  \item Morozov, S., McCairns, R. J. S., Merila, J. (2019). FishResp: R package and GUI application for analysis of aquatic respirometry data. Conservation Physiology, 7(1), coz003.
  \item Clark, T. D., Sandblom, E., Jutfelt, F. (2013). Aerobic scope measurements of fishes in an era of climate change: respirometry, relevance and recommendations. Journal of Experimental Biology, 216(15), 2771-2782.
  \item Svendsen, M. B. S., Bushnell, P. G., Steffensen, J. F. (2016). Design and setup of intermittent-flow respirometry system for aquatic organisms. Journal of Fish Biology, 88(1), 26-50.
  }
}
}
\author{
Sergey Morozov \email{sergey.morozov@helsinki.fi} Scott McCairns \email{scott.mccairns@inra.fr}
}
