% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases_functions.R
\name{fix_phase_gaps}
\alias{fix_phase_gaps}
\title{Fix phase gaps}
\usage{
fix_phase_gaps(input)
}
\arguments{
\item{input}{A list containing imported experiment data.
The output from \code{\link{load_experiment}}}
}
\value{
The input list with a corrected phases list.
}
\description{
Automatically expands the end of phases so they end
just before the start of the next ones.
}
