% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases_functions.R
\name{fix_phase_overlaps}
\alias{fix_phase_overlaps}
\title{Fix phase overlaps}
\usage{
fix_phase_overlaps(input)
}
\arguments{
\item{input}{A list containing imported experiment data.
The output from \code{\link{load_experiment}}.}
}
\value{
The input list with a corrected phases list.
}
\description{
Automatically trims the end of phases so they do not
overlap with the beginning of the next ones.
}
