% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o2_calcs_functions.R
\name{fraction_mr}
\alias{fraction_mr}
\title{Break down MR during a single cycle}
\usage{
fraction_mr(input, probe, phase, smoothing)
}
\arguments{
\item{input}{The output of process_pyro_mr}

\item{probe}{Which probe to select}

\item{phase}{which phase to select}

\item{smoothing}{How many points should be gathered for each calculation of the rolling MR}
}
\description{
Allows determining how oxygen consumption evolved throughout a cycle. Particularly useful for
cycles performed after exercise, where metabolic rate might significantly decrease throughout
the measurement cycle. Calculates a linear model for each combination of points.
}
