% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyroscience.aquaresp.R
\name{load.aquaresp.file}
\alias{load.aquaresp.file}
\title{Load a file containing summary phase data obtained from the 'AquaResp' software}
\usage{
load.aquaresp.file(file, date.format)
}
\arguments{
\item{file}{The name of a file which contains summary data obtained from the 'AquaResp' software (\href{www.aquaresp.com}{free software})}

\item{date.format}{A string indicating date format used in the input (e.g. "%Y-%m-%d")}
}
\value{
A dataframe containing the summary data on the recorded phases
}
\description{
Load a file containing summary phase data obtained from the 'AquaResp' software
}
