% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.pyroscience_functions.R
\name{load.pyroscience.logger.file}
\alias{load.pyroscience.logger.file}
\title{Load a respirometry data file originating from pyroscience's 'Pyro Oxygen Logger' software.}
\usage{
load.pyroscience.logger.file(file, n.chamber = 1:4, date.format)
}
\arguments{
\item{file}{The name of a file which contains raw data obtained from the 'Pyro Oxygen Logger' software (\href{https://www.pyro-science.com}{PyroScience})}

\item{n.chamber}{integer: the number of chambers used in an experiment (including empty ones)}

\item{date.format}{string: date format used in raw data obtained from the 'Pyro Oxygen Logger' software (e.g. "\%Y-\%m-\%d")}
}
\value{
A dataframe containing the recorded oxygen data
}
\description{
Load a respirometry data file originating from pyroscience's 'Pyro Oxygen Logger' software.
}
