% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyroscience_functions.R
\name{load.pyroscience.pH.file}
\alias{load.pyroscience.pH.file}
\title{Load a single channel file}
\usage{
load.pyroscience.pH.file(file, date.format, skip = 0, tz = Sys.timezone())
}
\arguments{
\item{file}{the input file}
}
\value{
the imported channel file
}
\description{
Load a single channel file
}
