% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.pyroscience_functions.R
\name{load.pyroscience.workbench.file}
\alias{load.pyroscience.workbench.file}
\title{Load a respirometry data file originating from pyroscience's 'PyroScience Workbench' software.}
\usage{
load.pyroscience.workbench.file(file, date.format)
}
\arguments{
\item{file}{The name of a file which contains raw data obtained from the 'PyroScience Workbench' software (\href{https://www.pyro-science.com}{PyroScience})}

\item{date.format}{A string indicating date format used in raw data obtained from the 'Pyro Oxygen Logger' software (e.g. "\%Y-\%m-\%d")}
}
\value{
A dataframe containing the recorded oxygen data
}
\description{
Load a respirometry data file originating from pyroscience's 'PyroScience Workbench' software.
}
