% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{load_experiment}
\alias{load_experiment}
\title{Wrapper to load the data for an experiment run (pyro and phases files).}
\usage{
load_experiment(
  folder,
  date_format,
  tz = Sys.timezone(),
  phases,
  probe_info,
  encoding = "ISO-8859-1",
  mass_unit = "g",
  vol_unit = "ml"
)
}
\arguments{
\item{folder}{the path to the folder where the experiment data is stored}

\item{date_format}{the format used in the raw date values (locale dependent)}

\item{tz}{The time zone of the data. Defaults to the system time zone.}

\item{phases}{a phases object. The output of \code{\link{process_phases}}}

\item{probe_info}{a dataframe containing animal information. This
dataframe must contain the following columns:
- animal_id    : The ID of the animal
- animal_mass  : The mass of the animal, in grams
- animal_vol   : Optional: The volume of the animal.
Assumed to be equal to mass if missing.
- chamber_vol  : The non-corrected volume of the chamber
- probe        : The device-channel combination for the probe
- first_cycle  : The first cycle of valid data for that animal}

\item{encoding}{the encoding of the source file. Defaults to "iso-8859-1",
which is the encoding used by the pyro workbench software.}
}
\value{
A list containing a phases dataframe and a pyro list with the
individual source data frames (in source_data), as well as a single,
combined data frame organized by time (in compiled_data).
}
\description{
Scans the target folder for a phases file and the raw pyroscience
experiment files. Imports them.
}
