% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coolterm_functions.R
\name{load_legacy_coolterm_file}
\alias{load_legacy_coolterm_file}
\title{import flush and measurement phase times recorded by the program CoolTerm.}
\usage{
load_legacy_coolterm_file(file, max.gap.fix = 1)
}
\arguments{
\item{file}{The name of the file to be imported.}
}
\value{
A data.frame containing the start and stop timestamps of the recorded phases.
}
\description{
import flush and measurement phase times recorded by the program CoolTerm.
}
