% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases_functions.R
\name{load_phases}
\alias{load_phases}
\title{import phase times object from RData file created by Cricket}
\usage{
load_phases(file)
}
\arguments{
\item{file}{The name of the file to be imported.}
}
\value{
A data frame containing the phase information.
}
\description{
import phase times object from RData file created by Cricket
}
