% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyro_functions.R
\name{load_pyro_files}
\alias{load_pyro_files}
\title{dummy documentation}
\usage{
load_pyro_files(
  folder,
  legacy.phases = FALSE,
  legacy.device.name = TRUE,
  max.gap.fix = 1
)
}
\description{
scan a target folder for a phases file and a pyroscience experiment file. import them.
}
