% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyro_functions.R
\name{load_pyro_pH_file}
\alias{load_pyro_pH_file}
\title{Load a single channel file}
\usage{
load_pyro_pH_file(file, date.format, skip = 0, tz = Sys.timezone())
}
\arguments{
\item{file}{the input file}
}
\value{
the imported channel file
}
\description{
Load a single channel file
}
