% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt_resp.R
\name{melt_resp}
\alias{melt_resp}
\title{Convert a human-friendly respirometry table to a computer-friendly format}
\usage{
melt_resp(input)
}
\arguments{
\item{input}{A resp list containing a phased object.}
}
\description{
Also appends additional probe info if provided.
}
