% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases_functions.R
\name{merge_legacy_pyroscience_phases}
\alias{merge_legacy_pyroscience_phases}
\title{Merge the oxygen data obtained from the pyroscience software with the phases recorded by Coolterm.}
\usage{
merge_legacy_pyroscience_phases(pyroscience, phases)
}
\arguments{
\item{pyroscience}{The dataframe containing the O2 measurements and imported using \code{\link{load.pyroscience.logger.file}} or \code{\link{load.pyroscience.workbench.file}}.}

\item{phases}{The dataframe containing the phase information, as imported by the function \code{\link{load_wide_phases_file}}}
}
\value{
a data.frame similar to the pyroscience input, but where the Phase column has been updated
}
\description{
Merge the oxygen data obtained from the pyroscience software with the phases recorded by Coolterm.
}
