% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases_functions.R
\name{merge_pyro_phases}
\alias{merge_pyro_phases}
\title{Merge the pyro data with the phases.}
\usage{
merge_pyro_phases(input)
}
\arguments{
\item{input}{A list containing imported experiment data.
The output from \code{\link{load_experiment}}}
}
\value{
An updated experiment list containing a phased data frame
}
\description{
Matches the two data sources by probe and timestamp and assigns the
respective phases to each datapoint. The phases object must not have
gaps/overlaps/repeated phase names for downstream functions to work
properly.
}
