% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o2_calcs_functions.R
\name{override_bg}
\alias{override_bg}
\title{Override background correction factor for any given
chambers using the values from another chamber.}
\usage{
override_bg(bg, replace, with)
}
\description{
Override background correction factor for any given
chambers using the values from another chamber.
}
