% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyro_functions.R
\name{patch_NAs}
\alias{patch_NAs}
\title{Fill in missing datapoints using the nearest available data.}
\usage{
patch_NAs(input, na_action = c("linear", "before", "after"), verbose = TRUE)
}
\arguments{
\item{input}{A dataframe containing Timestamps on the first column and
matching data on the remaining. If input contains a "phase" column, it will
be transported to the output unchanged.}

\item{na_action}{One of:
'linear' to capture the nearest before and after non-NA values and make a
linear interpolation.
'before' to find the nearest non-NA value before the NA and use it to fill
the gap.
'after'  to do the same as above but with the nearest value coming after
the NA.}

\item{verbose}{Logical. Should exceptions be reported out loud.
Defaults to TRUE.}
}
\value{
The input table with the NAs filled in as requested.
}
\description{
At high sampling rates, data points can occasionally be lost, causing odd
interruptions in the temperature and/or oxygen traces.
This function fills in those gaps.
}
