% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_bg}
\alias{plot_bg}
\title{Plot background readings}
\usage{
plot_bg(input, probes, linewidth = 1.5)
}
\arguments{
\item{input}{An experiment list with calculated background.
The output of \code{\link{calc_bg}}.}

\item{probes}{A string of which probes to plot}

\item{linewidth}{The width of the averaged/linear bg.}
}
\value{
a ggplot object
}
\description{
Plot background readings
}
