% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_deltas}
\alias{plot_deltas}
\title{Plot the calculated oxygen deltas}
\usage{
plot_deltas(input, cycles, probes, verbose = TRUE)
}
\arguments{
\item{input}{An experiment list with calculated deltas.
The output of \code{\link{process_experiment}} or
\code{\link{calc_delta}}.}

\item{cycles}{A numeric vector of which cycles to plot}

\item{probes}{A string of which probes to plot}

\item{verbose}{should argument warnings be displayed?}
}
\value{
a ggplot object
}
\description{
Plot the calculated oxygen deltas
}
