% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_meas}
\alias{plot_meas}
\title{Plot raw measurements of oxygen and temperature}
\usage{
plot_meas(
  input,
  cycles,
  chambers,
  temperature = FALSE,
  oxygen.label = "Oxygen"
)
}
\arguments{
\item{input}{A dataframe of measurements imported using one of FishResp's loading functions}

\item{temperature}{Logical: Should the temperature recordings be plotted?}

\item{oxygen.label}{The label of the Oxygen axis. Defaults to "Oxygen"}

\item{phases}{A vector of phases to plot (leave empty to plot all phases)}
}
\value{
A ggplot object
}
\description{
Plot raw measurements of oxygen and temperature
}
