% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_mr}
\alias{plot_mr}
\title{Dummy documentation}
\usage{
plot_mr(
  MR,
  SMR = NULL,
  MMR = NULL,
  cycles,
  chambers,
  target_col = "MR.mass",
  ylabel = expression(paste("MR (", mg ~ O[2] ~ kg^-1 ~ h^-1, ")"))
)
}
\description{
Dummy documentation
}
