% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_rolling_mmr}
\alias{plot_rolling_mmr}
\title{Plot rolling mmr}
\usage{
plot_rolling_mmr(rolling_mmr)
}
\arguments{
\item{rolling_mmr}{the output of compile_rolling_mmr (which can have been edited by  exclude_rolling_mmr_segment)}
}
\description{
Plot rolling mmr
}
