% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_slopes}
\alias{plot_slopes}
\title{Plot the calculated slopes}
\usage{
plot_slopes(input, cycles, probes, r2 = TRUE, verbose = TRUE)
}
\arguments{
\item{input}{An experiment list with calculated slopes.
The output of \code{\link{process_experiment}} or
\code{\link{calc_slopes}}.}

\item{cycles}{A numeric vector of which cycles to plot}

\item{probes}{A string of which probes to plot}

\item{r2}{show the respective R2 for each slope?}

\item{verbose}{should argument warnings be displayed?}
}
\value{
a ggplot object
}
\description{
Plot the calculated slopes
}
