% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_smr}
\alias{plot_smr}
\title{Plot the standard metabolic rates}
\usage{
plot_smr(input)
}
\arguments{
\item{input}{An experiment list with calculated SMR.
The output of \code{\link{process_experiment}} or
\code{\link{calc_smr}}.}
}
\value{
a ggplot object
}
\description{
Plot the standard metabolic rates
}
