% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{process_mr}
\alias{process_mr}
\title{Wrapper to perform metabolic rate calculations}
\usage{
process_mr(input, r2 = 0.95, G = 1:4, q = c(0.2, 0.25), p = 0.1, n = 10)
}
\arguments{
\item{input}{The output of \code{\link{subtract_bg}}}

\item{r2}{numeric: minimal coefficient of determination (\eqn{r^{2}})
for valid slopes. Default \eqn{r^{2}} = 0.95.}

\item{G}{A vector of the number of clusters to use for MLND SMR estimations.
Defaults to 1:4. Set to NULL to skip MLND method.}

\item{q}{A vector of the quantiles to use for the quantile SMR estimations.
Defaults to 0.2 and 0.25. Set to NULL to skip the quantile method.}

\item{p}{A vector of the percentages to use for low percent SMR estimations.
Percent methods discard the five lowest measurements before calculating the
average of the lowest x percentage. Defaults to 0.1. Set to NULL to skip
the percent method.}

\item{n}{A vector of the number of measurements to use for the low SMR
estimations. This method grabs the lowest n measurements and averages them.
No values are discarded using this method. Defaults to 10. Set to NULL to
skip the low n method.}
}
\value{
an updated input list containing the following new objects:
\itemize{
\item \code{all_slopes}: The slopes calculated for each cycle.
\item \code{good_slopes}: The slopes which pass the R2 threshold.
\item \code{mr}: The metabolic rates calculated from the good slopes.
\item \code{smr}: A data frame with the different SMR metrics calculated.
Relevant details for the different methods are saved in the attributes.
\item \code{mmr}: A data frame containing the cycle with the highest
metabolic rate recorded.
}
}
\description{
Calculates the slopes, filters by threshold R2, calculates metabolic rate for
each cycle, and from there calculates various SMR metrics and extracts MMR.
}
