% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases_functions.R
\name{process_phases}
\alias{process_phases}
\title{process the phase data}
\usage{
process_phases(phases, tz = Sys.timezone())
}
\arguments{
\item{phases}{a phases data.frame. The output of \code{\link{load_phases}}}

\item{tz}{the timezone of the timestamp data. Defaults to the system timezone.}
}
\value{
an updated phases object
}
\description{
Calculates start and stop times for each phase
}
