% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{process_probe_info}
\alias{process_probe_info}
\title{needs a better place in the future.
needs to be exported in the future.}
\usage{
process_probe_info(input, vol_unit = "ml", mass_unit = "g")
}
\description{
needs a better place in the future.
needs to be exported in the future.
}
\keyword{internal}
