% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyro_functions.R
\name{process_pyro_files}
\alias{process_pyro_files}
\title{dummy documentation}
\usage{
process_pyro_files(
  input,
  wait,
  chamber.info,
  O2_unit,
  min_temp,
  max_temp,
  start_time,
  stop_time,
  patch_NAs = TRUE
)
}
\description{
perform standard processing operations to the pyro/phases files
}
