% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyroscience_functions.R
\name{process_pyro_mr}
\alias{process_pyro_mr}
\title{dummy documentation}
\usage{
process_pyro_mr(
  input,
  r2,
  O2_raw,
  smr.method = "calcSMR.low10pc",
  max.length = 99999
)
}
\description{
perform standard metabolic rate calculations in pyro datasets
}
