% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyroresp.R
\docType{package}
\name{pyroresp}
\alias{pyroresp}
\alias{pyroresp-package}
\title{pyroresp: Analyse respirometry data captured by firesting loggers}
\description{
Designed to analyse intermittent-flow respirometry data. This is a sibling package of the FishResp R package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hugomflavio/pyroresp}
  \item Report bugs at \url{https://github.com/hugomflavio/pyroresp/issues}
}

}
\author{
\strong{Maintainer}: Hugo Flávio \email{hflavio@wlu.ca} (\href{https://orcid.org/0000-0002-5174-1197}{ORCID})

}
