% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyroscience.aquaresp.R
\name{pyroscience.aquaresp}
\alias{pyroscience.aquaresp}
\title{Convert Respirometry Data from PyroScience and AquaResp Software to the FishResp Format}
\usage{
pyroscience.aquaresp(
  input,
  pyro.type = c("logger", "workbench"),
  aquaresp,
  n.chamber = 1:4,
  date.format,
  wait.phase = NA,
  measure.phase = NA
)
}
\arguments{
\item{pyroscience.file}{the name of a file which contains raw data obtained from the 'Pyro Oxygen Logger' software (\href{https://www.pyroscience.com/}{PyroScience})}

\item{aquaresp.file}{the name of a file which contains summary data obtained from the 'AquaResp' software (\href{https://www.aquaresp.com}{free software})}

\item{aquaresp}{a dataframe or name of a file which contains summary data obtained from the 'AquaResp' software (\href{www.aquaresp.com}{free software})}

\item{n.chamber}{integer: the number of chambers used in an experiment (including empty ones)}

\item{date.format}{string: date format used in raw data obtained from the 'Pyro Oxygen Logger' software (e.g. "\%Y-\%m-\%d")}

\item{wait.phase}{integer: duration of the wait phase (in seconds), see the 'AquaResp' summary file (row #5)}

\item{measure.phase}{integer: duration of the measure phase (in seconds), see the 'AquaResp' summary file (row #6)}
}
\value{
The function returns a data frame containing raw data in the 'FishResp' format
}
\description{
The function is used to convert raw data from 'Pyro Oxygen Logger' (\href{https://www.pyroscience.com/}{PyroScience}) and a summary file from 'AquaResp' (\href{https://www.aquaresp.com}{free software}) to 'FishResp' format. This function should be applied before usage of the functions \code{\link{import.test}} and \code{\link{import.meas}}. The output is a file containing raw respirometry data in the 'FishResp' format (see Details in \code{\link{import.test}} to read more information about the 'FishResp' format)
}
\examples{
\dontrun{
pyroscience.path = system.file("extdata/pyroscience/pyroscience.txt",
                 package = "FishResp")
aquaresp.path = system.file("extdata/pyroscience/pyroscience-aquaresp.txt",
                 package = "FishResp")

pyroscience.aquaresp(file = pyroscience.path,
                     pyro.type = "logger",
                     aquaresp = aquaresp.path,
                     date.format = "\%m/\%d/\%Y",
                     n.chamber = 1,
                     wait.phase = 120,
                     measure.phase = 600)
}
}
