% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pyro_functions.R
\name{read_pyro_raw_file}
\alias{read_pyro_raw_file}
\title{Load a single raw channel file}
\usage{
read_pyro_raw_file(
  file,
  date_format,
  skip = 0,
  tz = Sys.timezone(),
  encoding = "iso-8859-1"
)
}
\arguments{
\item{file}{the path to a raw channel data file.}

\item{date_format}{the format used in the raw date values (locale dependent)}

\item{skip}{Number of data rows to skip. Defaults to 0. Note: This is not the
number of lines to skip from the top of the file. It is the number of lines
to skip once the actual raw data starts.}

\item{tz}{The time zone of the data. Defaults to the system time zone.}

\item{encoding}{the encoding of the source file. Defaults to "iso-8859-1",
which is the encoding used by the pyro workbench software.}
}
\value{
A data frame containing the inported data
}
\description{
Load a single raw channel file
}
