% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases_functions.R
\name{rename_phases}
\alias{rename_phases}
\title{Rename phases from scratch}
\usage{
rename_phases(input)
}
\arguments{
\item{input}{A list containing imported experiment data.
The output from \code{\link{load_experiment}}}
}
\value{
The input list with a corrected phases list.
}
\description{
Replaces recorded phase names with new ones automatically starting from
M1 and F1. If the first phase is a flush, then that one gets renamed
to F0 instead.
}
