% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_functions.R
\name{replace_bg}
\alias{replace_bg}
\title{Replace background readings for any given
probes using the values from another probe.}
\usage{
replace_bg(input, replace, with)
}
\arguments{
\item{input}{The output of \code{\link{calc_bg}}.}

\item{replace}{A string of probe names to replace.}

\item{with}{The probe to use as a reference for replacement.}
}
\value{
the updated bg dataframe
}
\description{
Replace background readings for any given
probes using the values from another probe.
}
