% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases_functions.R
\name{replicate_phases}
\alias{replicate_phases}
\title{Expand phases to other chambers}
\usage{
replicate_phases(input, from_device, to_device, from_channel, to_channel)
}
\arguments{
\item{input}{The phases list}

\item{from_device}{Name of the device from which to copy}

\item{to_device}{Name of the device to which to copy (can be the same)}

\item{from_channel}{Name of the channel within "from_device" to duplicate}

\item{to_channel}{Name of the channel (or channels if vector) to which to duplicate.}
}
\description{
Useful when using one flush pump for many chambers
}
