% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_functions.R
\name{roll_mr}
\alias{roll_mr}
\title{Break down MR from a single cycle}
\usage{
roll_mr(input, probe, cycle, smoothing, density = 1, r2 = 0.95)
}
\arguments{
\item{input}{The output of \code{\link{process_mr}}}

\item{probe}{Which probe to select}

\item{cycle}{which cycle to select}

\item{smoothing}{How many points should be gathered for
each calculation of the rolling MR}

\item{density}{The density. defaults to 1 (1 g = 1 ml).
Higher densities lead to lower volumes, e.g. if density = 2,
then 1 g = 0.5 ml.}

\item{r2}{Minimal coefficient of determination (\eqn{r^{2}})
for valid slopes. Defaults to \eqn{r^{2}} = 0.95.}
}
\description{
Allows determining how oxygen consumption evolved throughout a cycle.
Particularly useful for cycles performed after exercise, where metabolic
rate might significantly decrease throughout the measurement cycle.
Calculates a linear model for each combination of points.
}
