% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{sem}
\alias{sem}
\title{calculate standard error of the mean}
\usage{
sem(x, na.rm = TRUE)
}
\arguments{
\item{x}{a vector of numbers}

\item{na.rm}{Defaults to TRUE}
}
\value{
the SEM value
}
\description{
calculate standard error of the mean
}
