% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_functions.R
\name{subtract_bg}
\alias{subtract_bg}
\title{Subtract background from slopes}
\usage{
subtract_bg(
  input,
  pre,
  post,
  method = c("pre", "post", "average", "linear", "exponential", "parallel", "none"),
  ref_probe
)
}
\arguments{
\item{input}{A list containing cleaned oxygen data.
Obtained through \code{\link{process_experiment}}.}

\item{pre}{(optional) A data frame containing pre-test background readings.
Obtained through \code{\link{calc_bg}}.}

\item{post}{(optional) A data frame containing post-test background readings.
Obtained through \code{\link{calc_bg}}.}

\item{method}{the name of the method used for background correction:
#' \itemize{
\item  "pre" - subtracts the pre-experiment background readings
from the experiment oxygen consumption readings.
\item  "post" - subtracts the post-experiment background readings
from the experiment oxygen consumption readings.
\item  "average" - averages the pre- and post-experiment background readings
and subtracts them from the experiment oxygen consumption readings.
\item  "linear" - calculates a linear progression between pre- and
post-experiment background readings and subtracts the respective
background from the experiment oxygen consumption readings, using
the phase number as an indicator of the position in the linear
progression.
\item  "exponential" - calculates an exponential progression between pre- and
post-experiment background readings and subtracts the respective
background from the experiment oxygen consumption readings, using
the phase number as an indicator of the position in the linear
progression.
\item  "parallel" - subtracts the oxygen consumption readings of one probe
(listed using ref_probe) from the remaining probes, matching both
by cycle.
\item  "none" - does not perform oxygen consumption subtraction.
Not recommended for anything other than checking test data.
}}

\item{ref_probe}{string: the name of an empty probe used
only for the method 'parallel'}
}
\value{
A data frame containing data of metabolic rate measurements
corrected for background respiration.
}
\description{
Subtract background from slopes
}
\references{
{Svendsen, M. B. S., Bushnell, P. G., & Steffensen, J. F. (2016).
Design and setup of intermittent-flow respirometry system for aquatic
organisms. Journal of Fish Biology, 88(1), 26-50.}
}
