% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{suppress_EOL_warning}
\alias{suppress_EOL_warning}
\title{suppress incomplete final line warning}
\usage{
suppress_EOL_warning(expr)
}
\arguments{
\item{expr}{the read.table function call}
}
\value{
The outcome of the expression
}
\description{
Data files do not end with an EOL. This is by design, so a warning should
not be thrown. This function catches that specific warning and suppresses
it, while letting other warnings be shown.
}
\keyword{internal}
