% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{transfer_attributes}
\alias{transfer_attributes}
\title{helper function to avoid attribute loss
caused by the split->lapply->rbind process}
\usage{
transfer_attributes(from, to)
}
\arguments{
\item{from}{The original data frame}

\item{to}{The remade data frame}
}
\value{
"to", with any new attributes transferred from "from."
}
\description{
helper function to avoid attribute loss
caused by the split->lapply->rbind process
}
\keyword{internal}
