% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o2_calcs_functions.R
\name{trim_resp}
\alias{trim_resp}
\title{Prepare imported measurements for further analyses}
\usage{
trim_resp(
  input,
  wait = 0,
  cycle_max = Inf,
  auto_cut_last = FALSE,
  first_cycle = 1
)
}
\arguments{
\item{input}{the data frame containing imported oxygen measurements}

\item{wait}{integer: the number of first rows for each measurement phase (M)
which should be reassigned to the wait phase (W). Note: If your
phase-tracking device already assigns a wait phase, set this to 0.}

\item{cycle_max}{integer: Max allowed length (in number of rows) that each
cycle is allowed to have. Defaults to Inf (i.e. the whole cycle is used for
metabolic rate estimations).}

\item{auto_cut_last}{Should the last recorded phase be automatically cut?
useful if the experiment was terminated mid-measurement.}

\item{first_cycle}{either a vector of length one with the first cycle for all
probes, or a named vector containing the first cycle for each individual
probe (with the respective probe name).}
}
\value{
A data frame containing measurements valid for further analyses
}
\description{
Prepare imported measurements for further analyses
}
