% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coolterm_functions.R
\name{update_live_coolterm}
\alias{update_live_coolterm}
\title{import flush and measurement phase times recorded by the program CoolTerm.}
\usage{
update_live_coolterm(coolterm, max.gap.fix = Inf)
}
\arguments{
\item{file}{The name of the file to be imported.}
}
\value{
A data.frame containing the start and stop timestamps of the recorded phases.
}
\description{
Updated function for the new pump controller with four individual controls
}
