% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coolterm_functions.R
\name{update_phase_merge}
\alias{update_phase_merge}
\title{Merge the oxygen data obtained from the pyroscience software with the phases recorded by Coolterm.}
\usage{
update_phase_merge(pyro, coolterm)
}
\arguments{
\item{coolterm}{The dataframe containing the phase information, as imported by the function \code{\link{load_wide_coolterm_file}}}

\item{pyroscience}{The dataframe containing the O2 measurements and imported using \code{\link{load.pyroscience.logger.file}} or \code{\link{load.pyroscience.workbench.file}}.}
}
\value{
a data.frame similar to the pyroscience input, but where the Phase column has been updated
}
\description{
updated to work with the four channel pump controller
}
