% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_functions.R
\name{assign_device_names}
\alias{assign_device_names}
\title{Assign device names after completing experiment}
\usage{
assign_device_names(folder, assign_list)
}
\arguments{
\item{folder}{the path to the folder where the experiment data is stored}

\item{assign_list}{A list of format device_letter = device_name for the
devices to rename based on their letter.}
}
\value{
Nothing. Used for side effects.
}
\description{
To use if you forgot to assign a device name at the start of the experiment.
Goes through the raw data files, renames the required device names, and
resaves the files.
}
\details{
NOTE: This function will modify the files in your data folder!
}
