% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg_functions.R
\name{calc_bg}
\alias{calc_bg}
\title{Calculate background values}
\usage{
calc_bg(input, method = c("mean", "first", "last"))
}
\arguments{
\item{input}{The output of \code{\link{process_experiment}}.}

\item{method}{How should multiple background cycles for a single probe
be handled? One of:
'mean'  : All the measured cycles are used for the calculations
'first' : Only the first cycle is used for the calculations
'last'  : Only the last cycle is used for the calculations}
}
\value{
The input list, with a new "bg" object.
}
\description{
Calculate background values
}
