% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_functions.R
\name{calc_mr}
\alias{calc_mr}
\title{Calculate metabolic rates}
\usage{
calc_mr(slope_data, density = 1)
}
\arguments{
\item{slope_data}{a data frame obtained by using either the
function \code{\link{calc_slopes}} or \code{\link{filter_r2}}}

\item{density}{The density. defaults to 1 (1 g = 1 ml).
Higher densities lead to lower volumes, e.g. if density = 2,
then 1 g = 0.5 ml.}
}
\value{
A data frame with mr_abs, bg, and mr_cor.
}
\description{
Calculates the absolute and mass-specific metabolic rates,
as well as the relative importance of background (in percentage).
}
