% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_functions.R
\name{calc_slopes}
\alias{calc_slopes}
\title{Calculate slopes from clean measurements}
\usage{
calc_slopes(input)
}
\arguments{
\item{input}{A dataframe of clean measurements.
The output of \code{\link{trim_resp}}.}
}
\description{
Calculate slopes from clean measurements
}
