% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epoc_functions.R
\name{extract_epoc}
\alias{extract_epoc}
\title{Extracts the cumulative AUC}
\usage{
extract_epoc(input, max_duration = Inf)
}
\arguments{
\item{input}{a resp object with an "auc" sub-object.
The output of \code{\link{calc_auc}}}

\item{max_duration}{maximum time allowed for the auc to reach 0 (in hours).
if AUC does not reach 0 before this threshold, the cumulative AUC for the
whole duration is used.}
}
\value{
the input object, with an 'epoc' sub-object
}
\description{
Until the stepwise AUC reaches 0, or max_duration is reached.
}
