% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/killen_summary.R
\name{killen_summary}
\alias{killen_summary}
\title{Compile a list of points following Killen et al. 2021}
\usage{
killen_summary(input)
}
\value{
a list of summary information
}
\description{
For more information on these points, see Killen et al. 2021:
}
\references{
Killen, S. S., Christensen, E. A. F., Cortese, D., Závorka, L.,
Norin, T., Cotgrove, L., Crespel, A., Munson, A., Nati, J. J. H.,
Papatheodoulou, M., & McKenzie, D. J. (2021). Guidelines for reporting
methods to estimate metabolic rates by aquatic intermittent-flow
respirometry. Journal of Experimental Biology, 224(18), jeb242522.
\url{https://doi.org/10.1242/jeb.242522}
}
