% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{load_experiment}
\alias{load_experiment}
\title{Wrapper to load the data for an experiment run (pyro and phases files).}
\usage{
load_experiment(
  folder,
  date_format,
  tz = Sys.timezone(),
  phases_file = "CoolTerm",
  probe_info,
  fix_phases = TRUE
)
}
\arguments{
\item{folder}{the path to the folder where the experiment data is stored}

\item{date_format}{the format used in the raw date values (locale dependent)}

\item{tz}{The time zone of the data. Defaults to the system time zone.}

\item{phases_file}{a pattern to match for a phases file inside the folder}

\item{probe_info}{a dataframe containing animal information. This
dataframe must contain the following columns:
- id             : The ID of the animal
- mass           : The mass of the animal, in grams
- volume         : The non-corrected volume of the chamber + tubing
- probe          : The device-channel combination for the_probe
- first_cycle : The first cycle of valid data for that animal}

\item{fix_phases}{Automatically correct any overlaps, gaps, or repeated
phase names found in the file.}
}
\value{
A list containing a phases dataframe and a pyro dataframe.
}
\description{
Scans the target folder for a phases file and the raw pyroscience
experiment files. Imports them.
}
