% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{load_pyro_data}
\alias{load_pyro_data}
\title{Wrapper to scan a pyro folder and load raw data files.}
\usage{
load_pyro_data(folder, date_format, tz)
}
\arguments{
\item{folder}{the pyroscience run folder,
containing a "ChannelData" folder inside}

\item{date_format}{the format used in the raw date values (locale dependent)}

\item{tz}{The time zone of the data. Defaults to the system time zone.}
}
\description{
Wrapper to scan a pyro folder and load raw data files.
}
