% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt_resp.R
\name{melt_resp}
\alias{melt_resp}
\title{Convert a human-friendly respirometry table to a computer-friendly format}
\usage{
melt_resp(input, probe_info)
}
\arguments{
\item{input}{A dataframe with oxygen, temperature, pressure, and phase data.
The output of \code{\link{merge_pyro_phases}}}

\item{probe_info}{a dataframe containing animal information. This
dataframe must contain the following columns:
- id             : The ID of the animal
- mass           : The mass of the animal, in grams
- volume         : The non-corrected volume of the chamber + tubing
- probe          : The device-channel combination for the_probe
- first_cycle : The first cycle of valid data for that animal}
}
\description{
Also appends additional probe info if provided.
}
