% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_experiment}
\alias{plot_experiment}
\title{dummy documentation}
\usage{
plot_experiment(
  pre,
  post,
  mr,
  cycles,
  chamber,
  smr = FALSE,
  mmr = FALSE,
  raw_delta_col = "O2.delta.raw",
  title = "Experiment measurements",
  mr.col = "MR.mass",
  mr.label = "MO2",
  verbose = FALSE
)
}
\description{
plots the whole experiment data, from background to MO2
}
