% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_meas}
\alias{plot_meas}
\title{Plot raw measurements of oxygen and temp}
\usage{
plot_meas(
  input,
  cycles,
  probes,
  type = c("lines", "points", "both"),
  show_temp = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{input}{An experiment list with melted oxygen and temperature data.
The output of \code{\link{process_experiment}} or
\code{\link{melt_resp}}.}

\item{cycles}{A numeric vector of which cycles to plot}

\item{probes}{A string of which probes to plot}

\item{type}{Should the plot show only "lines", "points", or "both"?}

\item{show_temp}{Should the temperature recordings be plotted?}

\item{verbose}{should argument warnings be displayed?}
}
\value{
A ggplot object
}
\description{
Plot raw measurements of oxygen and temp
}
