% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{process_experiment}
\alias{process_experiment}
\title{Wrapper to get experiment data ready for further analyses}
\usage{
process_experiment(
  input,
  wait,
  meas_max = Inf,
  meas_min = 60,
  original_o2,
  convert_o2_to,
  na_action = c("ignore", "linear", "before", "after", "remove"),
  zero_buffer = 3,
  first_cycle = 1,
  min_temp,
  max_temp,
  start_time,
  stop_time,
  from_cycle,
  to_cycle,
  verbose = TRUE
)
}
\arguments{
\item{input}{The output of \code{\link{load_experiment}}}

\item{meas_max}{integer: Max number of seconds to keep from each
measurement phase. Defaults to Inf (i.e. the whole phase is used for
metabolic rate estimations).}

\item{meas_min}{integer: Minimum number of seconds a measurement phase must
have to be considered valid. Defaults to 10.}

\item{convert_o2_to}{The o2 unit desired for the final results}

\item{zero_buffer}{when calculating the delta, a value must be assigned to 0.
Traditionally, this is the very first value of the cycle. However, due to
natural probe noise, this can cause an upward or downard shift to the whole
delta line. This can be countered by calculating the mean of a few initial
values, instead of relying only on the very first. zero_buffer sets how many
values should be used to estimate the starting O2 concentration for the
cycle.}

\item{first_cycle}{either a vector of length one with the first cycle for all
probes, or a named vector containing the first cycle for each individual
probe (with the respective probe name).}

\item{min_temp, max_temp}{For temperature ramp experiments. The minimum OR maximum temperatures
that must be reached before data is considered valid. Discards all phases
prior to this temperature being reached. Use only one of the two arguments
at a time.}

\item{start_time, stop_time}{Trim the experiment to a specific time period. You may use one or both of
these arguments at the same time. Input must be a string in
YYYY-MM-DD HH:MM:SS format.}

\item{from_cycle, to_cycle}{Trim the experiment to a specific group of cycles. You may use one or both
of these arguments at the same time. Input must be numeric.}

\item{verbose}{Logical. Should steps being taken be detailed with messages.
Defaults to TRUE.}

\item{patch_NAs}{Logical. Should NA values found in the raw data be patched?
Defaults to TRUE.}
}
\value{
An updated experiment list, containing a cleaned object with the
processed data.
}
\description{
Perform standard processing operations to the pyro/phases files.
}
