% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{process_experiment}
\alias{process_experiment}
\title{Wrapper to get experiment data ready for further analyses}
\usage{
process_experiment(
  input,
  wait,
  convert_o2_unit_to,
  min_temp,
  max_temp,
  start_time,
  stop_time,
  from_cycle,
  to_cycle,
  patch_NAs = TRUE,
  patch_method = c("linear", "before", "after"),
  verbose = TRUE
)
}
\arguments{
\item{input}{The output of \code{\link{load_experiment}}}

\item{wait}{The number of seconds to discard as wait phase}

\item{convert_o2_unit_to}{The o2 unit desired for the final results}

\item{min_temp, max_temp}{For temperature ramp experiments. The minimum OR maximum temperatures
that must be reached before data is considered valid. Discards all phases
prior to this temperature being reached. Use only one of the two arguments
at a time.}

\item{start_time, stop_time}{Trim the experiment to a specific time period. You may use one or both of
these arguments at the same time. Input must be a string in
YYYY-MM-DD HH:MM:SS format.}

\item{from_cycle, to_cycle}{Trim the experiment to a specific group of cycles. You may use one or both
of these arguments at the same time. Input must be numeric.}

\item{patch_NAs}{Logical. Should NA values found in the raw data be patched?
Defaults to TRUE.}

\item{patch_method}{One of:
'linear' to capture the nearest before and after non-NA values and make a
linear interpolation.
'before' to find the nearest non-NA value before the NA and use it to fill
the gap.
'after'  to do the same as above but with the nearest value coming after
the NA.}

\item{verbose}{Logical. Should steps being taken be detailed with messages.
Defaults to TRUE.}
}
\value{
An updated experiment list, containing a cleaned object with the
processed data.
}
\description{
Perform standard processing operations to the pyro/phases files.
}
