% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{process_slopes}
\alias{process_slopes}
\title{Wrapper to get calculate slopes, correct them, and filter them}
\usage{
process_slopes(input, r2 = 0.95, pre, post, method, correct_for_volume = FALSE)
}
\arguments{
\item{input}{The output of \code{\link{process_experiment}}}

\item{r2}{numeric: minimal coefficient of determination (\eqn{r^{2}})
for valid slopes. Default \eqn{r^{2}} = 0.95.}

\item{pre}{(optional) A data frame containing pre-test background readings.
Obtained through \code{\link{calc_bg}}.}

\item{post}{(optional) A data frame containing post-test background readings.
Obtained through \code{\link{calc_bg}}.}

\item{method}{the name of the method used for background correction:
#' \itemize{
\item  "pre" - subtracts the pre-experiment background readings
from the experiment oxygen consumption readings.
\item  "post" - subtracts the post-experiment background readings
from the experiment oxygen consumption readings.
\item  "average" - averages the pre- and post-experiment background readings
and subtracts them from the experiment oxygen consumption readings.
\item  "linear" - calculates a linear progression between pre- and
post-experiment background readings and subtracts the respective
background from the experiment oxygen consumption readings, using
the phase number as an indicator of the position in the linear
progression.
\item  "parallel" - subtracts the oxygen consumption readings of one probe
(listed using ref_probe) from the remaining probes, matching both
by cycle.
\item  "none" - does not perform oxygen consumption subtraction.
Not recommended for anything other than checking test data.
}}

\item{correct_for_volume}{Logical: Should the corrected chamber volume be
used to remove the volume units from the slope? Defaults to FALSE, which
is the right choice in most cases. If you are using a chamber to correct
the background of other chambers but they don't have the same volume
(not recommended), then you can set this to TRUE to minimise the impacts
of the different volumes on the results. Note that chambers with different
formats will have different surface-to-volume relationships, and might
therefore not be good controls of each other.}
}
\value{
An updated experiment list, containing two new objects; one with
all the slopes, and another with the slopes that pass the r2 threshold.
}
\description{
Wrapper to get calculate slopes, correct them, and filter them
}
