% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phases_functions.R
\name{replicate_phases}
\alias{replicate_phases}
\title{Replicate phases from one probe to others}
\usage{
replicate_phases(input, from_device, from_channel, to_device, to_channel)
}
\arguments{
\item{input}{A list containing imported experiment data.
The output from \code{\link{load_experiment}}}

\item{from_device}{Name of the device from which to copy}

\item{from_channel}{Name of the channel within "from_device" to duplicate}

\item{to_device}{Name of the device to which to copy (can be the same)}

\item{to_channel}{Name of the channel (or channels if vector)
to which to replicate.}
}
\value{
The input list with a corrected phases list.
}
\description{
Useful when using one flush pump for many chambers/probes
}
