% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mr_functions.R
\name{roll_mr}
\alias{roll_mr}
\title{Break down MR from a single cycle}
\usage{
roll_mr(input, probe, phase, smoothing, d = 1)
}
\arguments{
\item{input}{The output of \code{\link{process_mr}}}

\item{probe}{Which probe to select}

\item{phase}{which phase to select}

\item{smoothing}{How many points should be gathered for
each calculation of the rolling MR}

\item{d}{The density. defaults to 1 (1g = 1ml). Defaults to 1.
Higher densities lead to lower volumes, e.g. if d = 2, then 1g = 0.5ml.}
}
\description{
Allows determining how oxygen consumption evolved throughout a cycle.
Particularly useful for cycles performed after exercise, where metabolic
rate might significantly decrease throughout the measurement cycle.
Calculates a linear model for each combination of points.
}
