% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o2_calcs_functions.R
\name{trim_resp}
\alias{trim_resp}
\title{Prepare imported measurements for further analyses}
\usage{
trim_resp(
  input,
  meas_max = Inf,
  meas_min = 60,
  cut_last = FALSE,
  first_cycle = 1
)
}
\arguments{
\item{input}{the data frame containing imported oxygen measurements}

\item{meas_max}{integer: Max number of seconds to keep from each
measurement phase. Defaults to Inf (i.e. the whole phase is used for
metabolic rate estimations).}

\item{meas_min}{integer: Minimum number of seconds a measurement phase must
have to be considered valid. Defaults to 10.}

\item{cut_last}{Should the last recorded phase be automatically cut?
useful if the experiment was terminated mid-measurement.}

\item{first_cycle}{either a vector of length one with the first cycle for all
probes, or a named vector containing the first cycle for each individual
probe (with the respective probe name).}
}
\value{
A data frame containing measurements valid for further analyses
}
\description{
Prepare imported measurements for further analyses
}
